Add pydocstyle configuration
A new pydocstyle configuration file is introduced for Google convention, helping in maintaining and enhancing the standards of the Python docstrings in the project. The added rules contain specifications for different aspects of docstring such as whitespaces, line breaks, docstring sections, etc. Additional exception handling is added to some functions in marimba.py for better error handling and readability.