Refactor pre-commit hooks and update linting config
In this update, the definition of pre-commit hooks has been streamlined for better readability and maintenance. Each hook has been set up to run locally using the poetry environment. A pytest.ini file has also been added in the project to designate the test directories for pytest. The .pylintrc has been updated to ignore linting the tests and temp directories.