Add concurrency module to Marimba Standard Library
A new module ‘concurrency’ has been added to the Marimba Standard Library. This module includes parallel functionality, such as multithreaded generation of thumbnails from a list of images. It makes use of a decorator for multithreaded execution and ensures thread safety when appending to the list of generated thumbnail paths.