Implement hard links in package command
The 'copy' boolean flag has been replaced with an 'operation' enum to allow for more file operation modes. This update has been applied across several modules in the marimba project. Now, when dealing with files, we no longer only have the options to copy or move, but also to create a link, providing greater flexibility.